app/requirements.txt
# APi Server
fastapi
uvicorn[standard]
pydantic

# Database
motor
dnspython

python-dotenv